refactor(follow): CursorBasedPaginatedResult 도메인 계층으로 이동
#7
#11
