+++ src/main/java/com/gorogoro/followandlike/follow/infrastructure/jpa/FollowRepositoryJpaAdapter.java
@@ -1,6 +1,7 @@
 package com.gorogoro.followandlike.follow.infrastructure.jpa;
 
 import com.gorogoro.followandlike.follow.domain.model.Follow;
+import com.gorogoro.followandlike.follow.domain.model.dto.CursorBasedPaginatedResult;
 import com.gorogoro.followandlike.follow.domain.repository.FollowRepository;
 import lombok.NonNull;
 import org.jspecify.annotations.Nullable;
